+++ opencti-platform/opencti-graphql/src/modules/indicator/indicator.ts
@@ -27,12 +27,75 @@ const INDICATOR_DEFINITION: ModuleDefinition<StoreEntityIndicator, StixIndicator
       },
     },
   },
-  attributes: [ // FIXME
+  attributes: [
     { name: 'name', type: 'string', mandatoryType: 'external', editDefault: true, multiple: false, upsert: true },
     { name: 'description', type: 'string', mandatoryType: 'customizable', editDefault: true, multiple: false, upsert: true },
-    { name: 'content', type: 'string', mandatoryType: 'customizable', editDefault: true, multiple: false, upsert: true },
-    { name: 'content_mapping', type: 'string', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true },
-    { name: 'context', type: 'string', mandatoryType: 'external', editDefault: true, multiple: false, upsert: true },
+    {
+      name: 'pattern_type',
+      type: 'string',
+      mandatoryType: 'external',
+      editDefault: true,
+      multiple: false,
+      upsert: false,
+      label: 'Pattern type'
+    },
+    { name: 'pattern_version', type: 'string', mandatoryType: 'no', editDefault: false, multiple: false, upsert: false, label: 'Pattern version' },
+    { name: 'pattern', type: 'string', mandatoryType: 'external', editDefault: true, multiple: false, upsert: false },
+    {
+      name: 'indicator_types',
+      type: 'string',
+      mandatoryType: 'customizable',
+      editDefault: true,
+      multiple: true,
+      upsert: true,
+      label: 'Indicator types'
+    },
+    {
+      name: 'valid_from',
+      type: 'date',
+      mandatoryType: 'customizable',
+      editDefault: true,
+      multiple: false,
+      upsert: true,
+      label: 'Valid from'
+    },
+    {
+      name: 'valid_until',
+      type: 'date',
+      mandatoryType: 'customizable',
+      editDefault: true,
+      multiple: false,
+      upsert: true,
+      label: 'Valid until'
+    },
+    {
+      name: 'x_opencti_score',
+      type: 'numeric',
+      mandatoryType: 'customizable',
+      editDefault: true,
+      multiple: false,
+      upsert: true,
+      label: 'Score'
+    },
+    { name: 'x_opencti_detection', type: 'boolean', mandatoryType: 'no', editDefault: false, multiple: false, upsert: true },
+    {
+      name: 'x_opencti_main_observable_type',
+      type: 'string',
+      mandatoryType: 'external',
+      editDefault: true,
+      multiple: false,
+      upsert: true,
+      label: 'Main observable type'
+    },
+    {
+      name: 'x_mitre_platforms',
+      type: 'string',
+      mandatoryType: 'customizable',
+      editDefault: true,
+      multiple: true,
+      upsert: true,
+      label: 'Platforms'
+    },
   ],
   relations: [],
   representative: (stix: StixIndicator) => {
